Clean post cache after updating menu_order on featured listings
After running $wpdb->update to change the menu_order for featured
listings, we need to clean the post cache, similar to how this is
handled in wp_update_post and wp_insert_post. If the post cache isn't
cleaned, the menu_order of -1 is not saved to the database.

Removed added_post_meta action hook for maybe_update_menu_order. This is
not necessary as the first step in creating a job listing is to set
_featured to 0. All subsequent post meta changes for _featured are
therefore updates.