Change language and working of "single job type" to multi type.
Fixes #549
The multiselect category checkbox on the "Job Listings" settings page is similar in functionality to the job types one so I'm moving it to the "Job Listings" settings tab and renaming the single job type chekbox to "Multi-select Listing Types". The forms and related function had to be changed too.