Add link to date format to avoid confusion
Fixes #697
The Italian and French translations are incorrect as the translators changed the date form from "yyyy" and "dd" to the first letter of the words in those languages. I'm adding a link to the official documentation for the date format in PHP and Javascript respectively for each of these to help translation.