Warn the user if they're editing an existing job.
See #835

The plugin allows the user to edit jobs before they're published, but if the paid listings plugin is installed a new job may be in "pending_payment" status even after the user has submitted it. This patch warns the user and displays a link   to delete the cookie. The link will allow users to start a new job submission even without the paid listing plugin installed.