Warn the user if they're editing an existing job. (#847)
* Warn the user if they're editing an existing job.

See #835

The plugin allows the user to edit jobs before they're published, but if the paid listings plugin is installed a new job may be in "pending_payment" status even after the user has submitted it. This patch warns the user and displays a link   to delete the cookie. The link will allow users to start a new job submission even without the paid listing plugin installed.

* Check for paid listings before job listing while resuming and move code

If we're using the paid listings plugin and it shows packages before
jobs are displayed then the job_id will be set and resuming using the
cookie doesn't work properly.
Moved cookie reset code into the abstract form class so it's neater,
maybe.