Fix preview and editing conflicts
We updated from REQUEST to GET because if the user has a pending preview on cookies and try to edit another job, we were setting a GET of editing job and the POST of the preview (that comes from the cookie). In this case we were always getting from the POST because of the EGPCS order (https://www.php.net/manual/en/ini.core.php#ini.variables-order).

Also we had to update the preview validation to get the cookies only when we don't have a job_id from the GET (editing cases).