Disable cache when the order is set to random or random featured
This guarantees that the set of jobs displayed is always random. The bad aspect is, of course, that it will require more requests to the database just to get the list of jobs, but the alternative here would be to get a big pool of jobs from the database and limit that on the PHP side, but guess what? That's not very efficient either, nor allow it to show a random set of the full pool of jobs registered.

As users on #1967 were already disabling cache for this use case, I think it's worth to just do the same here, but only when ordering randomly.

By the way, this also disables the shuffling of cached results, as those are not required anymore because the result set will always be randomly ordered by the database itself.