Add support for WP.com Marketplace Addon purchase flow
When an addon is purchase through the WP.com Marketplace, the new integration would automatically configure the addon with the licensing information.

This is done by using a WPCOM specific filter that's executed after the provision_license event is triggered. This filter contains the licensing information needed by the plugin to connect to WPJobManager.com back-end API.