Update `$company_tagline` check for PHP8+
As of PHP8, passing `null` to `strlen()` is deprecated. The `$company_tagline` var is assigned to the output of the `get_the_company_tagline` function, which returns `null` under certain conditions.

With this change, we're checking that `$company_tagline` exists: continue if it does, but if it's `null` or doesn't exist, return early.