Handle case where base64_encode returns only invalid characters
Although rare, there's a scenario where base64_encode can return a sting composed only by the removed characters: '+', '/', '='. For instance,"+/++/+/++++/++/+++++++++/++++++///////////////////////+/+++++/+/==" is a valid Base 64 string.

Given that we there's the risk that random_bytes would return such a string (because it is, well, random). I decided to add a fallback so the string is converted to hexadecimal in cases where the base64 - after having all these characters removed - is empty.